# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  /api/v1/users:
    put:
      tags:
        - Управление пользователями
      summary: Изменение данных пользователя
      operationId: updateUser
      requestBody:
        description: Запрос на изменение данных пользователя
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
            examples:
              Запрос на изменение данных пользователя:
                description: Запрос на изменение данных пользователя
                value:
                  requestId: 123e4567-e89b-12d3-a456-426614174000
                  userId: 1231232134
                  username: NewNick
                  roles:
                    - name: USER
        required: true
      responses:
        '200':
          description: Успешная обработка запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                Ответ с данными пользователя:
                  description: Ответ с данными пользователя
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    userInfoList:
                      - userId: 123233241
                        username: Nick
                        firstName: Joe
                        lastName: Doe
                        roles:
                          - name: USER
        '400':
          description: >-
            Запрос не соответствует заявленной схеме данных и не может быть
            обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке валидации:
                  description: Ответ с сообщением об ошибке валидации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Нарушение формата данных
        '403':
          description: Обработка закончилась с ошибкой авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке авторизации:
                  description: Ответ с сообщением об ошибке авторизации
                  value:
                    operationStatus: ERROR
                    responseId: 91a0a5a8-9428-46b2-961e-d14a8485b48b
                    errorMessage: 'Ошибка авторизации: Access Denied'
        '404':
          description: Обработка закончилась с ошибкой поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке поиска:
                  description: Ответ с сообщением об ошибке поиска
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Пользователь не найден
        '500':
          description: Обработка закончилась внутренней ошибкой сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением о внутренней ошибкой сервиса:
                  description: Ответ с сообщением о внутренней ошибкой сервиса
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Внутренняя ошибка сервиса
      security:
        - JWT: []
    post:
      tags:
        - Управление пользователями
      summary: Создание пользователя
      operationId: createUser
      requestBody:
        description: Запрос на создание пользователя
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
            examples:
              Запрос на создание пользователя:
                description: Запрос на создание пользователя
                value:
                  requestId: 123e4567-e89b-12d3-a456-426614174000
                  username: Nick
                  password: password
                  firstName: Joe
                  lastName: Doe
                  roles:
                    - name: USER
        required: true
      responses:
        '200':
          description: Успешная обработка запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                Ответ с данными пользователя:
                  description: Ответ с данными пользователя
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    userInfoList:
                      - userId: 123233241
                        username: Nick
                        firstName: Joe
                        lastName: Doe
                        roles:
                          - name: USER
        '400':
          description: >-
            Запрос не соответствует заявленной схеме данных и не может быть
            обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке валидации:
                  description: Ответ с сообщением об ошибке валидации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Нарушение формата данных
        '403':
          description: Обработка закончилась с ошибкой авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке авторизации:
                  description: Ответ с сообщением об ошибке авторизации
                  value:
                    operationStatus: ERROR
                    responseId: 91a0a5a8-9428-46b2-961e-d14a8485b48b
                    errorMessage: 'Ошибка авторизации: Access Denied'
        '404':
          description: Обработка закончилась с ошибкой поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке поиска:
                  description: Ответ с сообщением об ошибке поиска
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Пользователь не найден
        '500':
          description: Обработка закончилась внутренней ошибкой сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением о внутренней ошибкой сервиса:
                  description: Ответ с сообщением о внутренней ошибкой сервиса
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Внутренняя ошибка сервиса
      security:
        - JWT: []
    delete:
      tags:
        - Управление пользователями
      summary: Удаление данных пользователя
      operationId: deleteUser
      requestBody:
        description: Запрос на удаление данных пользователя
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
            examples:
              Запрос на удаление данных пользователя:
                description: Запрос на удаление данных пользователя
                value:
                  requestId: 123e4567-e89b-12d3-a456-426614174000
                  userId: 1231232134
        required: true
      responses:
        '200':
          description: Успешная обработка запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                Ответ с данными пользователя:
                  description: Ответ с данными пользователя
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    userInfoList:
                      - userId: 123233241
                        username: Nick
                        firstName: Joe
                        lastName: Doe
                        roles:
                          - name: USER
        '400':
          description: >-
            Запрос не соответствует заявленной схеме данных и не может быть
            обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке валидации:
                  description: Ответ с сообщением об ошибке валидации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Нарушение формата данных
        '403':
          description: Обработка закончилась с ошибкой авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке авторизации:
                  description: Ответ с сообщением об ошибке авторизации
                  value:
                    operationStatus: ERROR
                    responseId: 91a0a5a8-9428-46b2-961e-d14a8485b48b
                    errorMessage: 'Ошибка авторизации: Access Denied'
        '404':
          description: Обработка закончилась с ошибкой поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке поиска:
                  description: Ответ с сообщением об ошибке поиска
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Пользователь не найден
        '500':
          description: Обработка закончилась внутренней ошибкой сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением о внутренней ошибкой сервиса:
                  description: Ответ с сообщением о внутренней ошибкой сервиса
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Внутренняя ошибка сервиса
      security:
        - JWT: []
  /api/v1/users/search:
    post:
      tags:
        - Управление пользователями
      summary: Поиск пользователей
      operationId: searchUser
      parameters:
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: direction
          in: query
          required: true
          schema:
            type: string
            enum:
              - ASC
              - DESC
        - name: sortField
          in: query
          required: true
          schema:
            type: string
      requestBody:
        description: Поиск пользователей с фильтрацией и пагинацией
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
            examples:
              Запрос на поиск пользователей:
                description: Запрос на поиск пользователей
                value:
                  requestId: 123e4567-e89b-12d3-a456-426614174000
                  userId: 1231232134
                  username: Nick
                  firstName: Joe
                  lastName: Doe
        required: true
      responses:
        '200':
          description: Успешная обработка запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                Ответ с данными пользователя:
                  description: Ответ с данными пользователя
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    userInfoList:
                      - userId: 123233241
                        username: Nick
                        firstName: Joe
                        lastName: Doe
                        roles:
                          - name: USER
        '400':
          description: >-
            Запрос не соответствует заявленной схеме данных и не может быть
            обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке валидации:
                  description: Ответ с сообщением об ошибке валидации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Нарушение формата данных
        '403':
          description: Обработка закончилась с ошибкой авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке авторизации:
                  description: Ответ с сообщением об ошибке авторизации
                  value:
                    operationStatus: ERROR
                    responseId: 91a0a5a8-9428-46b2-961e-d14a8485b48b
                    errorMessage: 'Ошибка авторизации: Access Denied'
        '404':
          description: Обработка закончилась с ошибкой поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке поиска:
                  description: Ответ с сообщением об ошибке поиска
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Пользователь не найден
        '500':
          description: Обработка закончилась внутренней ошибкой сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением о внутренней ошибкой сервиса:
                  description: Ответ с сообщением о внутренней ошибкой сервиса
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Внутренняя ошибка сервиса
      security:
        - JWT: []
...
components:
  schemas:
    RoleRequest:
      type: object
      properties:
        name:
          type: string
          enum:
            - ADMIN
            - USER
      required:
        - name
    UserRequest:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        userId:
          type: integer
          format: int64
        username:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleRequest'
          uniqueItems: true
    RoleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          enum:
            - ADMIN
            - USER
    UserInfo:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleResponse'
          uniqueItems: true
    UserResponse:
      type: object
      properties:
        operationStatus:
          type: string
          enum:
            - COMPLETED
            - ERROR
        responseId:
          type: string
        userInfoList:
          type: array
          items:
            $ref: '#/components/schemas/UserInfo'
    ErrorResponse:
      type: object
      properties:
        operationStatus:
          type: string
          enum:
            - COMPLETED
            - ERROR
        responseId:
          type: string
        errorMessage:
          type: string
...
  securitySchemes:
    JWT:
      type: http
      in: header
      scheme: bearer
      bearerFormat: 'Bearer '